remove unused operations
