+++ connector/functions/bucket.go
@@ -77,46 +77,3 @@ func ProcedureUpdateStorageBucket(ctx context.Context, state *types.State, args 
 
 	return true, nil
 }
-
-// FunctionStorageBucketPolicy gets access permissions on a bucket or a prefix.
-func FunctionStorageBucketPolicy(ctx context.Context, state *types.State, args *common.StorageBucketArguments) (string, error) {
-	return state.Storage.GetBucketPolicy(ctx, args)
-}
-
-// FunctionStorageBucketNotification gets notification configuration on a bucket.
-func FunctionStorageBucketNotification(ctx context.Context, state *types.State, args *common.StorageBucketArguments) (*common.NotificationConfig, error) {
-	return state.Storage.GetBucketNotification(ctx, args)
-}
-
-// ProcedureSetStorageBucketNotification sets a new notification configuration on a bucket.
-func ProcedureSetStorageBucketNotification(ctx context.Context, state *types.State, args *common.SetBucketNotificationArguments) (bool, error) {
-	if err := state.Storage.SetBucketNotification(ctx, args); err != nil {
-		return false, err
-	}
-
-	return true, nil
-}
-
-// ProcedureSetStorageBucketReplication sets replication configuration on a bucket. Role can be obtained by first defining the replication target on MinIO
-// to associate the source and destination buckets for replication with the replication endpoint.
-func ProcedureSetStorageBucketReplication(ctx context.Context, state *types.State, args *common.SetStorageBucketReplicationArguments) (bool, error) {
-	if err := state.Storage.SetBucketReplication(ctx, args); err != nil {
-		return false, err
-	}
-
-	return true, nil
-}
-
-// FunctionGetBucketReplication gets current replication config on a bucket.
-func FunctionStorageBucketReplication(ctx context.Context, state *types.State, args *common.StorageBucketArguments) (*common.StorageReplicationConfig, error) {
-	return state.Storage.GetBucketReplication(ctx, args)
-}
-
-// RemoveBucketReplication removes replication configuration on a bucket.
-func ProcedureRemoveStorageBucketReplication(ctx context.Context, state *types.State, args *common.StorageBucketArguments) (bool, error) {
-	if err := state.Storage.RemoveBucketReplication(ctx, args); err != nil {
-		return false, err
-	}
-
-	return true, nil
-}
