--- connector/testdata/01-setup/mutation/02-setStorageBucketNotification/expected.json
@@ -1,1 +1,0 @@
-{ "operation_results": [{ "result": true, "type": "procedure" }] }
