--- connector/testdata/01-setup/mutation/02-setStorageBucketNotification/request.json
@@ -1,13 +1,0 @@
-{
-  "collection_relationships": {},
-  "operations": [
-    {
-      "type": "procedure",
-      "name": "setStorageBucketNotification",
-      "arguments": {
-        "bucket": "minio-bucket-test",
-        "queueConfigurations": []
-      }
-    }
-  ]
-}
