--- connector/testdata/02-get/query/storageBucketNotification/expected.json
@@ -1,13 +1,0 @@
-[
-  {
-    "rows": [
-      {
-        "__value": {
-          "cloudFunctionConfigurations": [],
-          "queueConfigurations": [],
-          "topicConfigurations": []
-        }
-      }
-    ]
-  }
-]
