--- connector/testdata/02-get/query/storageBucketPolicy/expected.json
@@ -1,1 +1,0 @@
-[{ "rows": [{ "__value": "" }] }]
