--- connector/testdata/02-get/query/storageBucketPolicy/request.json
@@ -1,18 +1,0 @@
-{
-  "arguments": {
-    "bucket": {
-      "type": "literal",
-      "value": "minio-bucket-test"
-    }
-  },
-  "collection": "storageBucketPolicy",
-  "collection_relationships": {},
-  "query": {
-    "fields": {
-      "__value": {
-        "column": "__value",
-        "type": "column"
-      }
-    }
-  }
-}
