--- connector/testdata/02-get/query/storageBucketReplication/expected.json
@@ -1,1 +1,0 @@
-[{ "rows": [{ "__value": { "role": null, "rules": [] } }] }]
